+++ src/app/[locale]/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import { ThemeProvider } from "@/components/theme/theme-provider";
-import { Navbar } from "@/components/navbar";
+import { ThemeProvider } from "@/components/theme/ThemeProvider";
+import { Navbar } from "@/components/Navbar";
 import { I18nProviderClient } from "../../../locales/client";
 import { TransitionProvider } from "@/components/animation/TransitionProvider";
 import { Analytics } from "@vercel/analytics/react";
