+++ src/components/activity/SessionsTable.tsx
