+++ src/components/dialog/DeleteActivityDialog.tsx
